servo: Merge #5661 - Stop using int/uint in CEF (from Ms2ger:cef); r=jdm
Source-Repo: https://github.com/servo/servo
Source-Revision: c42c2ae9b57b5960ce617e5b4c0ab75959e46957

UltraBlame original commit: 9480d898333d4b1e533d537bc39eeac8a21e115f

+++ servo/ports/cef/browser.rs
@@ -86,7 +86,7 @@ pub struct ServoCefBrowser {
     
     pub callback_executed: Cell<bool>,
 
-    id: int,
+    id: isize,
     servo_browser: RefCell<ServoBrowser>,
     message_queue: RefCell<Vec<WindowEvent>>,
 }
